Add BounceTools and Composer.

+++ BounceTales/src/bouncetales/ext/rsc/ResourceType.java
@@ -2,9 +2,9 @@ package bouncetales.ext.rsc;
 
 public class ResourceType {
 
-	public static final int IMAGE = 2;
-	public static final int MIDI = 3;
-	public static final int STRINGS = 4;
-	public static final int LAYOUT = 5;
-	public static final int LEVEL = 8;
+	public static final byte IMAGE = 2;
+	public static final byte MIDI = 3;
+	public static final byte STRINGS = 4;
+	public static final byte LAYOUT = 5;
+	public static final byte LEVEL = 8;
 }
